Use global counters instead of UUIDs for some simple unique values
UUIDs are an easy way to solve this issue but I don't think they're
necessary?

These are at the moment the only reason that router-bridge are pulling
in piles and piles of deno code to support the
`crypto.getRandomValues()` API. We can save a lot of complexity by using
this.

Another alternative is to use something like `nanoid/non-secure`, which
has some theoretical possibility of collisions due to using
`Math.random()`.